move hardcodedResults to the right place
